+++ langwatch/src/server/background/workers/collector/evaluations.ts
@@ -7,22 +7,14 @@ import {
 } from "../../../../server/evaluations/preconditions";
 import type { CheckPreconditions } from "../../../../server/evaluations/types";
 import { createLogger } from "../../../../utils/logger/server";
-import { slugify } from "../../../../utils/slugify";
 import { prisma } from "../../../db";
 import type { ElasticSearchEvaluation, Span } from "../../../tracer/types";
 import { elasticSearchEvaluationSchema } from "../../../tracer/types.generated";
 import { scheduleEvaluation } from "../../queues/evaluationsQueue";
 import type { CollectorJob, EvaluationJob } from "../../types";
+import { evaluationNameAutoslug } from "./evaluationNameAutoslug";
 
 const _logger = createLogger("langwatch:workers:collector:evaluations");
-
-export const evaluationNameAutoslug = (name: string) => {
-  const autoslug = slugify(name || "unnamed", {
-    lower: true,
-    strict: true,
-  }).replace(/[^a-z0-9]/g, "_");
-  return `customeval_${autoslug}`;
-};
 
 export const mapEvaluations = (
   data: CollectorJob,